backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
transformers
fastapi==0.109.0
email-validator
numpy<2
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
python-multipart==0.0.6
feedparser
